Can Load multiple catalogs now
• Multiple catalogs can be loaded now. Requires a specific set up.
For the base game assets, the catalog must be in the runtime path.
Every other catalog should go in a folder along with the asset bundles.
The Addressables system MUST have one catalog file within the runtime path. This is a limitation of the system without a workaround, as far as I can tell.