Changed Addressables Profile Paths
• Changed paths to exclude BuildTarget.
It's a little clumsy, but there's not much I can do short of making my own changes to the addressables system to accomodate for multiple catalogs. Somethings will just have to be added by hand. If I had more time, I could devise an automated solution, but this will suffice.