Squashed Disjointed Tire Bug
The bug in question occurred when one of the sides of the car was not touching the ground. The reason being I was using world coords to calculate the position of the tire when I should've been using local coords AND the raycast hit returned {0,0,0} when the raycast hit nothing, causing the position of the tires to be incorrect. So, for example, if hit.collider = null (meaning the raycast didn't hit anything, I'd have a point value of {0,0,0}. ie: the world origin. Since I was setting the world position of the tires, the tires try to stick closer to y = 0 when the car is off the ground. However, changing it to local values resulted in the opposite problem, where the tires appeared to start flying above the car when the car was airborne. Again, this is because hit.point, when collider = null, was the world origin, so when converting this to local coords, the y value was a high value relative to the car's position.

Should also lerp the values smoothly so they don't snap awkwardly to the position, but for now, it's fine.