Added a power curve for acceleration
Speed is now dictated by an animation curve called "powerCurve".
It's going to take a while for me to understand just what the hell is going on with this to make it feel better. Lots of tweaking. And rolling is still happening, even after adding anti-rolling bars. It's gonna take a lot more to understand vehicle physics, more time than I have. For that reason, I'm just going to push this to main, but keep this issue open. Maybe in the future, I'll figure it out.